refactor(cli): extract card-lifecycle handlers into work_commands_lifecycle.rs (card c0bd865c phase 2 / 5aa9e780)
## What this PR lands

Phase 2 of the 5-way split of work_commands.rs. The card-lifecycle handlers — Open/Claimed/InProgress/Review/Merged/Closed state-machine — move to a sibling module:

  Extracted to work_commands_lifecycle.rs:
    - run_create / run_seed
    - run_claim / run_release / resolve_my_active_claim
    - run_heartbeat / run_update / run_state / run_close
    - cli_can_set_state_directly (state guard)
    - close_transition_allowed_from (state guard)
    - refusal_message (per-state actionable refusal text)

## Sequencing

Phase 1 (PR #1042) extracted the print/render group as the lowest-risk demo. Phase 2 is the heaviest extraction in the c0bd865c plan — 12 functions, ~310 LOC moved. Phases 3-5 (git / gh / review) follow and pull cross-module callees (spawn_claim_worktree, open_pr_and_link, auto_spawn_review_card) into their own modules.

Until those phases land, lifecycle reaches those callees via `crate::work_commands::fn_name` — they're pub(crate) so the cross-module path works, and visibility tightens back down when each one moves into its own sibling.

## File sizes

  work_commands.rs:         1823 → 1459 LOC (-364, including orphan-doc cleanup)
  work_commands_lifecycle.rs:   0 →  365 LOC (new)

## Public surface unchanged

  pub use crate::work_commands_lifecycle::{
      run_claim, run_close, run_create, run_heartbeat, run_release, run_seed,
      run_state, run_update,
  };
  pub(crate) use crate::work_commands_lifecycle::{
      cli_can_set_state_directly, close_transition_allowed_from, refusal_message,
  };

main.rs sees `work_commands::run_create(...)` like before. Tests in work_commands.rs explicitly import the moved items from work_commands_lifecycle so the test module compiles.

## Tests + lints

  cargo test work_commands:: → 19 pass
  cargo clippy -p airc-cli --bin airc -- -D warnings → clean
  cargo fmt --check → clean

The 19 tests are the existing work_commands suite — all pass against the moved functions (refusal_message_for_merged_carries_actionable_guidance and friends).

## Anchor

c0bd865c follow-up sequence. Phase 1 = render (#1042 shipped). Phase 2 = lifecycle (this PR). Phases 3-5 = git / gh / review still carded as c63a1b6e, c7bdf2df, a3ede6de.
